requirements.txt
Flask==2.3.4
requests==2.31.0
python-dotenv==1.1.1
flask-cors==3.0.10
